refactor spend endpoints
